chore: rename packages according to convention
