+++ README.md
@@ -3,11 +3,11 @@
 
 ### jmxClient tool
 - Given JMX url and port, continuously reads tick times to file. 
-  Call using: _java -jar jmxClient.jar < JMX Url > < JMX Port > < Out Folder > < Duration >_
+  Call using: _java -jar jmxClient-shaded.jar < JMX Url > < JMX Port > < Out Folder > < Duration >_
 
 ### jmxDummyServer
 - Creates a dummy server to continuously create random tick time values. Useful for testing jmxClient and
-JMX configuration. Call using: _java -jar jmxDummyServer.jar < Output to file > < Duration >_
+JMX configuration. Call using: _java -jar jmxDummyServer-shaded.jar < Output to file > < Duration >_
 
 ### Testing
 - First run the dummy server using the runDummyServer.sh utility, then the jmxClient using runClientTest.sh.
