+++ src/main/java/client/JmxClient.java
@@ -1,4 +1,4 @@
-package jmxClient;
+package client;
 
 import java.io.FileWriter;
 import java.text.MessageFormat;
