+++ src/test/java/client/JmxClientTest.java
@@ -1,4 +1,4 @@
-package jmxClient;
+package client;
 
 import static org.junit.jupiter.api.Assertions.*;
